requirements.txt
asgiref==3.5.2
certifi==2022.9.24
charset-normalizer==2.1.1
coreapi==2.3.3
coreschema==0.0.4
Django==4.1.3
djangorestframework==3.14.0
et-xmlfile==1.1.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
numpy==1.23.4
openpyxl==3.0.10
packaging==21.3
pandas==1.5.1
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.6
requests==2.28.1
rest-pandas==1.1.0
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
six==1.16.0
sqlparse==0.4.3
uritemplate==4.1.1
urllib3==1.26.12
